requirements.txt
djangorestframework==3.x.x
imdbpy
django-cors-headers